+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Rust edition transition [#303]
+- No tilde expansion in `maildir-dir` [#305]
 
 ## [0.5.6] - 2022-02-22
 
@@ -438,3 +439,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#288]: https://github.com/soywod/himalaya/issues/288
 [#289]: https://github.com/soywod/himalaya/issues/289
 [#303]: https://github.com/soywod/himalaya/issues/303
+[#305]: https://github.com/soywod/himalaya/issues/305