+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.7] - 2022-03-01
+
+### Added
+
+- Notmuch support [#57]
+
 ### Fixed
 
 - Build failure due to `imap` version [#303]
@@ -16,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [**BREAKING**] Replace `inbox-folder`, `sent-folder` and `draft-folder` by a generic hashmap `mailboxes`
+- Display short envelopes id for `maildir` and `notmuch` backends [#309]
 
 ## [0.5.6] - 2022-02-22
 
@@ -314,7 +321,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Password from command [#22]
 - Set up README [#20]
 
-[unreleased]: https://github.com/soywod/himalaya/compare/v0.5.6...HEAD
+[unreleased]: https://github.com/soywod/himalaya/compare/v0.5.7...HEAD
+[0.5.7]: https://github.com/soywod/himalaya/compare/v0.5.6...v0.5.7
 [0.5.6]: https://github.com/soywod/himalaya/compare/v0.5.5...v0.5.6
 [0.5.5]: https://github.com/soywod/himalaya/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/soywod/himalaya/compare/v0.5.3...v0.5.4
@@ -374,6 +382,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#48]: https://github.com/soywod/himalaya/issues/48
 [#50]: https://github.com/soywod/himalaya/issues/50
 [#54]: https://github.com/soywod/himalaya/issues/54
+[#57]: https://github.com/soywod/himalaya/issues/57
 [#58]: https://github.com/soywod/himalaya/issues/58
 [#59]: https://github.com/soywod/himalaya/issues/59
 [#61]: https://github.com/soywod/himalaya/issues/61
@@ -446,3 +455,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#303]: https://github.com/soywod/himalaya/issues/303
 [#305]: https://github.com/soywod/himalaya/issues/305
 [#308]: https://github.com/soywod/himalaya/issues/308
+[#309]: https://github.com/soywod/himalaya/issues/309